this work in now obsolete
there is some problems with the tree created as most parts of it are not needed.  shriniva has already fixed these problems in her branch perceptron which is better made without the problems that this one had.  this is only being updated just incase some parts of it may be useful later.